Added comments and small refactoring
