+++ src/components/GlobalStyle.tsx
@@ -1,3 +1,6 @@
+/**
+ * Global styles applied to the entire app
+ */
 import { createGlobalStyle } from "styled-components";
 
 export const GlobalStyle = createGlobalStyle`
@@ -8,7 +11,6 @@ export const GlobalStyle = createGlobalStyle`
       "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
       sans-serif;
     -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
   }
   code {
     font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
