+++ src/services/chromeService.ts
@@ -14,9 +14,6 @@ export const getBookmarks = async () => {
   }
 };
 
-export const openNewTab = (params: { url: string }) =>
-  new Promise<chrome.tabs.Tab>(resolve => chrome.tabs.create(params, resolve));
-
 export const moveBookmark = async (
   bookmark: Bookmark,
   oldIndex: number,
@@ -24,6 +21,10 @@ export const moveBookmark = async (
 ) => {
   if (NODE_ENV !== "development") {
     return new Promise<ChromeBookmark>(resolve => {
+      // The Chrome move API seems to have a bug when you move a bookmark to
+      // a position where the new index is greater than the current one.
+      // Increasing the updated index position by 1 for this specific case
+      // seems to fix the issue: https://stackoverflow.com/q/13264060/4836602
       let fixedNewIndex = newIndex;
       if (oldIndex < newIndex) {
         fixedNewIndex++;
